Add files via upload
Heads or Tails
In this exercise, we are going to focus on writing clean and organized code using functions! When using functions it is important to be descriptive and to organize your code to have a separation of concerns. Let's see this in action:

Create a function tossCoin() ->  String
Have this function print "Tossing a Coin!"
Next have the function randomly pick either "Heads" or "Tails"
Have the function print the result
Finally, have the function return the result
Now create another function tossMultipleCoins(Int) -> Double
Have this function call the tossCoin function multiple times based on the Integer input
Have the function return a Double that reflects the ratio of head toss to total toss