Add files via upload
Graphics Primitives
In Computer Graphics there is the idea of primitives, which are the simple shapes that all drawn objects are composed of. The most commonlyused of these primitives are the Point, Line, and Triangle. Using swift structs we are going to replicate these three different graphics primitive types.

The point of the assignment is to learn the Swift OOP syntax. If you find yourself spending a lot of time on mathematical logic, ask a cohort-mate for help.

Objective
Get comfortable creating structs, and referencing them in other structs.