Add files via upload
Assignment: Cold Call
Create an application that will pick out a random name from an Array of names every time the cold call button is clicked. We will be hard coding this Array of names in our View Controller. We will be using Auto Layout in this assignment. Make sure your application has behavior similar to the gif image below:

Assignment: ColdCall v.2
Let's add another feature to our previous application. Whenever we click on the Cold Call button, we will not only get a random name displayed on the top but also a random number from 1 to 5 inclusive in the center of the screen. If the number is 1 or 2, the number should have a red color, if the number is 3 or 4 then the number should have an orange color, and finally, if the number is 5 then the number should have a green color.